+++ smawg/__init__.py
@@ -4,9 +4,7 @@ See https://github.com/expurple/smawg for more info about the project.
 """
 
 from smawg._engine import Game, Hooks, InvalidAssets, validate
-from smawg._plugin_interface import (
-    Ability, Combo, Player, Race, Region, RulesViolation
-)
+from smawg.common import Ability, Combo, Player, Race, Region, RulesViolation
 
 __all__ = [
     "Game", "Hooks", "InvalidAssets", "validate",